Option to read file using system command - reject
Feature where user can optionally read file using a system command
instead of the internal readfile().

Idea was that this would prevent files from being added to the hidden
buflist – which is wrong. The files are added to the hidden bufflist
when one update the Quickfix list.

So this is dead. Leave it here in case some of the code is needed later.